DAT89: Made the pre read and write hooks act just like post create.
Why it was done?
At the moment, the agent will check with the server about each action the user
is trying to do on the file. Therefore, the flow should be the same -
read magic and fileId, give it to the user mode along side the action
(will be implemented) and then gets if it is safe for open or not.

In the future, when the magic&fileId parse will work, it is better to
combine all the functionality to one function and just call it for
each hook.
Right now it just a copy paste :)